Remove reliance on ng-app or data-ng-app for bootstrapping angular
The angular 2 documentation states that if you’re trying to bootstrap
a hybrid Angular 1/ Angular 2 application you should not bootstrap
angular with the ng-app tag. Rather, it recommends that you bootstrap
manually via the angular.bootstrap method.

Since this gem currently only looks for the ng-app tag, hybrid apps
will not work with this gem.

For a look at how protractor handles this, it’s a similar solution:

https://github.com/angular/protractor/blob/6a4dc7a6a5b796e0215e5b9abf99494ac13cb647/lib/clientsidescripts.js#L75